+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+* CXV1-3541 - Enabled assignment of message templates on users page
+
 ## [4.8.6]
 * Hide API keys configuration menu item